+++ dfm_tools/__init__.py
@@ -4,10 +4,9 @@
 
 __author__ = """Jelmer Veenstra"""
 __email__ = "jelmer.veenstra@deltares.nl"
-__version__ = "0.15.1"
+__version__ = "0.16.1"
 
 from dfm_tools.deprecated import *
-from dfm_tools.errors import *
 from dfm_tools.download import *
 from dfm_tools.get_nc import *
 from dfm_tools.get_nc_helpers import *
@@ -25,4 +24,14 @@ from dfm_tools import data
 from dfm_tools.modelbuilder import *
 
 import warnings
-warnings.filterwarnings("always",category=DeprecationWarning)
+warnings.filterwarnings(action="always", category=DeprecationWarning)
+
+# python 3.8 DeprecationWarning
+import sys
+version_info = sys.version_info
+if (version_info.major == 3) & (version_info.minor <= 8):
+    warnings.warn(
+        DeprecationWarning("You are using Python 3.8 or lower, please note that "
+                           "dfm_tools will soon drop support for Python 3.8: "
+                           "https://github.com/Deltares/dfm_tools/issues/267")
+        )