+++ dfm_tools/hydrolib_helpers.py
@@ -15,6 +15,23 @@ import datetime as dt
 import geopandas
 from shapely.geometry import LineString
 
+__all__ = ["Dataset_to_T3D",
+           "Dataset_to_TimeSeries",
+           "Dataset_to_Astronomic",
+           "DataFrame_to_PolyObject",
+           "geodataframe_to_PolyFile",
+           "DataFrame_to_TimModel",
+           "ForcingModel_to_plipointsDataset",
+           "forcinglike_to_Dataset",
+           "pointlike_to_DataFrame",
+           "TimModel_to_DataFrame",
+           "pointlike_to_geodataframe_points",
+           "PolyFile_to_geodataframe_points",
+           "PolyFile_to_geodataframe_linestrings",
+           "gdf_linestrings_to_points",
+           "tekalobject_to_DataFrame",
+       ]
+
 
 def get_ncbnd_construct():
     """
@@ -298,9 +315,7 @@ def ForcingModel_to_plipointsDataset(forcingmodel:hcdfm.ForcingModel, npoints=No
         plipointsDataset_list.append(ds_onepoint)
     ds = xr.concat(plipointsDataset_list, dim=dimn_point)
     
-    # prevent circular import #TODO: solve underlaying issue
-    from dfm_tools.interpolate_grid2bnd import _maybe_convert_fews_to_dfmt
-    ds = _maybe_convert_fews_to_dfmt(ds)
+    ds = maybe_convert_fews_to_dfmt(ds)
     
     return ds
 
@@ -527,7 +542,7 @@ def PolyFile_to_geodataframe_points(polyfile_object:hcdfm.PolyFile, crs:str=None
     return gdf
 
 
-def _da_from_gdf_points(gdf_points):
+def da_from_gdf_points(gdf_points):
     
     ncbnd_construct = get_ncbnd_construct()
     dimn_point = ncbnd_construct['dimn_point']
@@ -604,3 +619,46 @@ def tekalobject_to_DataFrame(polyobject):
     polyobject_pd.columns = polyobject.description.content.split('\n')[2:] #to fix legend labels
     polyobject_pd.index.name = polyobject.metadata.name
     return polyobject_pd
+
+
+def maybe_convert_fews_to_dfmt(ds):
+    ncbnd_construct = get_ncbnd_construct()
+    dimn_point = ncbnd_construct['dimn_point']
+    varn_pointname = ncbnd_construct['varn_pointname']
+    
+    # convert station data_vars to coords to avoid dfmt issues
+    for var_to_coord in [varn_pointname,'station_names']:
+        if var_to_coord in ds.data_vars:
+            ds = ds.set_coords(var_to_coord)
+    # assign timeseries_id cf_role to let FM read the station names
+    ds[varn_pointname] = ds[varn_pointname].assign_attrs({'cf_role': 'timeseries_id'})
+    
+    # rename data_vars to long_name (e.g. renames FEWS so to salinitybnd)
+    for datavar in ds.data_vars:
+        if datavar in ['ux','uy']: #TODO: keeping these is consistent with hardcoded behaviour in dfm_tools elsewhere, but not desireable
+            continue
+        if hasattr(ds[datavar],'long_name'):
+            longname = ds[datavar].attrs['long_name']
+            if longname.endswith('bnd'):
+                ds = ds.rename_vars({datavar:longname})
+    
+    # transpose dims #TODO: the order impacts the model results: https://issuetracker.deltares.nl/browse/UNST-7402
+    # dfmt (arbitrary) dimension ordering is node/time/z
+    # required to reorder to FEWS time/node/z order for comparable results
+    # also time/z/node will result in unexpected results
+    if "time" in ds.dims: # check if time dimension is present (astronomic does not have time)
+        ds = ds.transpose("time", dimn_point, ...)
+    
+    # convert station names to string format (keep attrs and encoding)
+    # also needed to properly export, since we cannot encode it at dtype S1 properly otherwise
+    if not ds[varn_pointname].dtype.str.startswith('<'):
+        with xr.set_options(keep_attrs=True):
+            ds[varn_pointname] = ds[varn_pointname].load().str.decode('utf-8',errors='ignore').str.strip() #.load() is essential to convert not only first letter of string.
+
+    # add relevant encoding if not present
+    ds[varn_pointname].encoding.update({'dtype': 'S1', 'char_dim_name': 'char_leng_id'})
+    
+    # assign global attributes #TODO: add more
+    ds = ds.assign_attrs({'Conventions': 'CF-1.6',
+                          'institution': 'Deltares'})
+    return ds