+++ tests/test_meshkernel_helpers.py
@@ -15,14 +15,16 @@ import numpy as np
 import geopandas as gpd
 from shapely.geometry import Polygon
 import glob
-from dfm_tools.meshkernel_helpers import _meshkernel_check_geographic, _geographic_to_meshkernel_projection
+from dfm_tools.meshkernel_helpers import (meshkernel_check_geographic, 
+                                          geographic_to_meshkernel_projection, 
+                                          add_crs_to_dataset)
 
 
 @pytest.mark.unittest
 def test_add_crs_to_dataset_cartesian():
     uds = xu.data.adh_san_diego()
     crs='EPSG:26946' # this is not the correct crs for this model, but that does not matter
-    dfmt.add_crs_to_dataset(uds,is_geographic=False,crs=crs)
+    add_crs_to_dataset(uds,is_geographic=False,crs=crs)
     
     assert 'projected_coordinate_system' in uds.data_vars
     crs_attrs = uds.projected_coordinate_system.attrs
@@ -35,14 +37,14 @@ def test_add_crs_to_dataset_cartesian():
 def test_add_crs_to_dataset_spherical():
     uds = xu.data.adh_san_diego()
     crs='EPSG:4326' # this is not the correct crs for this model, but that does not matter
-    dfmt.add_crs_to_dataset(uds,is_geographic=True,crs=crs)
+    add_crs_to_dataset(uds,is_geographic=True,crs=crs)
     
     assert 'wgs84' in uds.data_vars
     crs_attrs = uds.wgs84.attrs
     assert crs_attrs['EPSG_code'] == 'EPSG:4326'
     assert crs_attrs['epsg'] == 4326
     assert crs_attrs['grid_mapping_name'] == 'latitude_longitude'
-
+    
 
 @pytest.mark.unittest
 def test_meshkernel_delete_withcoastlines():
@@ -112,27 +114,44 @@ def test_meshkernel_delete_withgdf():
     assert len(mk.mesh2d_get().face_nodes) == 17368
 
 
-@pytest.mark.unittest
+@pytest.mark.systemtest
 def test_meshkernel_check_geographic():
+    """
+    to check whether is_geographic can be correctly derived from the mk object, for cartesian as well spherical objects
+    """
+    lon_min, lon_max, lat_min, lat_max = -68.55, -67.9, 11.8, 12.6
+    dxy = 0.05
+    
+    mk_cartesian = dfmt.make_basegrid(lon_min, lon_max, lat_min, lat_max, dx=dxy, dy=dxy, is_geographic=False)
+    mk_cartesian_geograph = meshkernel_check_geographic(mk_cartesian)
+    mk_spherical = dfmt.make_basegrid(lon_min, lon_max, lat_min, lat_max, dx=dxy, dy=dxy, is_geographic=True)
+    mk_spherical_geograph = meshkernel_check_geographic(mk_spherical)
+    
+    assert mk_cartesian_geograph==False
+    assert mk_spherical_geograph==True
+
+
+@pytest.mark.unittest
+def test_meshkernel_check_geographic2():
     
     mk = meshkernel.MeshKernel(projection=meshkernel.ProjectionType.CARTESIAN)
-    is_geographic = _meshkernel_check_geographic(mk)
+    is_geographic = meshkernel_check_geographic(mk)
     assert is_geographic==False
     
     mk = meshkernel.MeshKernel(projection=meshkernel.ProjectionType.SPHERICAL)
-    is_geographic = _meshkernel_check_geographic(mk)
+    is_geographic = meshkernel_check_geographic(mk)
     assert is_geographic==True
     
     mk = meshkernel.MeshKernel(projection=meshkernel.ProjectionType.SPHERICALACCURATE)
-    is_geographic = _meshkernel_check_geographic(mk)
+    is_geographic = meshkernel_check_geographic(mk)
     assert is_geographic==True
 
 
 @pytest.mark.unittest
 def test_geographic_to_meshkernel_projection():
     
-    spherical = _geographic_to_meshkernel_projection(is_geographic=True)
-    cartesian = _geographic_to_meshkernel_projection(is_geographic=False)
+    spherical = geographic_to_meshkernel_projection(is_geographic=True)
+    cartesian = geographic_to_meshkernel_projection(is_geographic=False)
     spherical_mk = meshkernel.ProjectionType.SPHERICAL
     cartesian_mk = meshkernel.ProjectionType.CARTESIAN