+++ docs/whats-new.md
@@ -1,6 +1,16 @@
 ## UNRELEASED
 
 ### Feat
+- interpolation of edge/node variables to faces with `dfmt.uda_to_faces()` (deprecates `dfmt.uda_edges_to_faces()`) by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#651](https://github.com/Deltares/dfm_tools/pull/651) and [#644](https://github.com/Deltares/dfm_tools/pull/644)
+
+### Fix
+- performance improvements of `dfmt.uda_to_faces()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#652](https://github.com/Deltares/dfm_tools/pull/652)
+
+
+## 0.16.0 (2023-11-03)
+
+### Feat
+- more robust support for CMEMS/CDS credentials including environment variables by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#633](https://github.com/Deltares/dfm_tools/issues/633)
 - enrich rst file with topology from corresponding mapfile with `dfmt.enrich_rst_with_map()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#626](https://github.com/Deltares/dfm_tools/issues/626)
 - add cellinfo to minimal 2D networks (with 1D topology) with `dfmt.add_network_cellinfo()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#624](https://github.com/Deltares/dfm_tools/issues/624)
 - xugrid feature `uds.ugrid.to_nonperiodic()` deprecates `dfmt.remove_periodic_cells()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#503](https://github.com/Deltares/dfm_tools/issues/503)
@@ -11,6 +21,7 @@
 - increased buffer in `dfmt.download_ERA5()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#612](https://github.com/Deltares/dfm_tools/pull/612)
 - support for Polygon geometries in `dfmt.geodataframe_to_PolyFile()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#610](https://github.com/Deltares/dfm_tools/pull/610)
 - fill nan-values in initial salinity/temperature netcdf dataset in `dfmt.preprocess_ini_cmems_to_nc()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#617](https://github.com/Deltares/dfm_tools/pull/617)
+- skip all-nan boundary support points instead of converting to zeros in `dfmt.plipointsDataset_to_ForcingModel()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#637](https://github.com/Deltares/dfm_tools/pull/637)
 
 
 ## 0.15.0 (2023-10-19)